Add "Type" column to DataGrid and update AssignmentInfo
A new "Type" column has been added to the DataGrid in the
AssignmentPage.xaml file, bound to the new `Type` property
in the `AssignmentInfo` class. The constructor of the
AssignmentPage class has been updated to include the `Type`
property for existing entries, all set to "Win32". A new
entry, "App Four", has been added with the properties:
Name = "App Four", Id = "004", Platform = "Windows", and
Type = "Win32".