Add AssignmentIntentComboBox to AssignmentPage.xaml
Introduced a new ComboBox named `AssignmentIntentComboBox` to the
`AssignmentPage.xaml`. This ComboBox is designed to allow users to
select an assignment intent, with options including "Required",
"Available", and "Uninstall". It features a header labeled
"Assignment Intent", a placeholder text "Select an intent", and
is set to a width of 150 with a default selected index of 0.