Add FilterToggle check for assignment filter type
Introduce a conditional check for the `FilterToggle` feature.
If `FilterToggle.IsOn` is true, set the filter type based on
`_selectedFilterMode` to either `Include` or `Exclude`. If
`FilterToggle.IsOn` is false, set the filter type to `None`.
This change allows the filter to be disabled when the toggle
is off, providing more control over filter application.